pyproject.toml
[project]
name = "agentic-binary-analysis"
version = "0.1.0"
description = "Agentic workflow for automated PE/ELF binary analysis"
requires-python = ">=3.11"

dependencies = [
  "agno",
  "fastmcp",
  "pefile",
  "pyelftools",
  "lief",
  "r2pipe",
  "google-generativeai",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
